+++ components/UserGroupAdminMethods.php
@@ -122,6 +122,24 @@ class UserGroupAdminMethods
         $this->consultation->refresh();
     }
 
+    public function setUserData(int $userId, string $nameGiven, string $nameFamily, string $organization, ?string $newPassword): void
+    {
+        $user = User::findOne(['id' => $userId]);
+        if (trim($nameGiven) === '' && trim($nameFamily) === '') {
+            return;
+        }
+
+        $user->nameFamily = trim($nameFamily);
+        $user->nameGiven = trim($nameGiven);
+        $user->organization = trim($organization);
+        $user->name = trim(trim($nameGiven) . ' ' . trim($nameFamily));
+        $user->save();
+
+        if ($newPassword !== null && trim($newPassword) !== '') {
+            $user->changePassword(trim($newPassword));
+        }
+    }
+
     private function getUserGroup(int $userGroupId): ?ConsultationUserGroup
     {
         foreach ($this->consultation->getAllAvailableUserGroups() as $userGroup) {