+++ models/db/User.php
@@ -40,7 +40,6 @@ use yii\web\IdentityInterface;
  * @property null|AmendmentSupporter[] $amendmentSupports
  * @property null|MotionComment[] $motionComments
  * @property null|MotionSupporter[] $motionSupports
- * @property Site[] $adminSites
  * @property ConsultationUserGroup[] $userGroups
  * @property ConsultationLog[] $logEntries
  * @property UserNotification[] $notifications
@@ -190,14 +189,28 @@ class User extends ActiveRecord implements IdentityInterface
         return $this->hasMany(ConsultationLog::class, ['userId' => 'id']);
     }
 
-    public function getAdminSites(): ActiveQuery
-    {
-        return $this->hasMany(Site::class, ['id' => 'siteId'])->viaTable('siteAdmin', ['userId' => 'id']);
-    }
-
     public function getUserGroups(): ActiveQuery
     {
         return $this->hasMany(ConsultationUserGroup::class, ['id' => 'groupId'])->viaTable('userGroup', ['userId' => 'id']);
+    }
+
+
+    private static array $preloadedConsultationUserGroups = [];
+    public static function preloadConsultationUserGroups(Consultation $consultation): void
+    {
+        if (isset(self::$preloadedConsultationUserGroups[$consultation->id])) {
+            return;
+        }
+
+        self::$preloadedConsultationUserGroups[$consultation->id] = [];
+        foreach ($consultation->getAllAvailableUserGroups([], true) as $userGroup) {
+            foreach ($userGroup->getUserIds() as $userId) {
+                if (!isset(self::$preloadedConsultationUserGroups[$consultation->id][$userId])) {
+                    self::$preloadedConsultationUserGroups[$consultation->id][$userId] = [];
+                }
+                self::$preloadedConsultationUserGroups[$consultation->id][$userId][] = $userGroup;
+            }
+        }
     }
 
     /**
@@ -205,6 +218,10 @@ class User extends ActiveRecord implements IdentityInterface
      */
     public function getConsultationUserGroups(Consultation $consultation): array
     {
+        if (isset(self::$preloadedConsultationUserGroups[$consultation->id][$this->id])) {
+            return self::$preloadedConsultationUserGroups[$consultation->id][$this->id];
+        }
+
         $groups = [];
         foreach ($consultation->getAllAvailableUserGroups([], true) as $userGroup) {
             foreach ($userGroup->getUserIds() as $userId) {
@@ -651,7 +668,7 @@ class User extends ActiveRecord implements IdentityInterface
      * @throws FormError
      * @throws \app\models\exceptions\ServerConfiguration
      */
-    public function sendRecoveryMail()
+    public function sendRecoveryMail(): void
     {
         if ($this->recoveryAt) {
             $recTs = Tools::dateSql2timestamp($this->recoveryAt);
@@ -678,10 +695,9 @@ class User extends ActiveRecord implements IdentityInterface
     }
 
     /**
-     * @param string $token
      * @throws FormError
      */
-    public function checkRecoveryToken($token): bool
+    public function checkRecoveryToken(string $token): bool
     {
         if ($this->recoveryAt) {
             $recTs = Tools::dateSql2timestamp($this->recoveryAt);
@@ -697,12 +713,7 @@ class User extends ActiveRecord implements IdentityInterface
         return true;
     }
 
-    /**
-     * @param string $newEmail
-     * @param int $timestamp
-     * @return string
-     */
-    public function createEmailChangeToken($newEmail, $timestamp): string
+    public function createEmailChangeToken(string $newEmail, int $timestamp): string
     {
         if (YII_ENV == 'test' && mb_strpos($newEmail, '@example.org') !== false) {
             return 'testCode';
@@ -713,11 +724,9 @@ class User extends ActiveRecord implements IdentityInterface
     }
 
     /**
-     * @param string $newEmail
-     * @param string $code
      * @throws FormError
      */
-    public function checkEmailChangeToken($newEmail, $code)
+    public function checkEmailChangeToken(string $newEmail, string $code): void
     {
         if ($this->emailChange != $newEmail || $this->emailChange === null) {
             throw new FormError(\Yii::t('user', 'err_emailchange_notfound'));
@@ -735,7 +744,7 @@ class User extends ActiveRecord implements IdentityInterface
      * @throws MailNotSent
      * @throws ServerConfiguration
      */
-    public function sendEmailChangeMail(string $newEmail)
+    public function sendEmailChangeMail(string $newEmail): void
     {
         $changeTs            = time();
         $this->emailChange   = $newEmail;
@@ -756,7 +765,7 @@ class User extends ActiveRecord implements IdentityInterface
     /**
      * @throws FormError
      */
-    public function changeEmailAddress(string $newEmail, string $code)
+    public function changeEmailAddress(string $newEmail, string $code): void
     {
         if (AntragsgruenApp::getInstance()->confirmEmailAddresses) {
             $this->checkEmailChangeToken($newEmail, $code);